extract parseForESLint
