+++ eslint-bridge/tests/rules/rule-test-launcher.test.ts
@@ -23,9 +23,7 @@ import { Rule, RuleTester } from 'eslint';
 
 import { rules } from 'rules/main';
 import { readAssertions } from '../testing-framework/assertions';
-import { buildSourceCode } from 'parser';
 import { readFileSync } from 'fs';
-import { FileType } from '../../src/analyzer';
 
 /**
  * Return test files for specific rule based on rule key
@@ -68,21 +66,7 @@ function runRuleTests(rules: Record<string, Rule.RuleModule>, ruleTester: RuleTe
   }
 }
 
-/**
- * This function is provided as 'parseForESLint' implementation which is used in RuleTester to invoke exactly same logic
- * as we use in our 'parser.ts' module
- */
-export function parseForESLint(
-  fileContent: string,
-  options: { filePath: string },
-  fileType: FileType = 'MAIN',
-) {
-  const { filePath } = options;
-  return buildSourceCode(
-    { filePath, fileContent, fileType, tsConfigs: [] },
-    filePath.endsWith('.ts') ? 'ts' : 'js',
-  );
-}
+
 
 const ruleTester = new RuleTester({ parser: __filename });
 runRuleTests(rules, ruleTester);
